Shuffle code to match tags
